% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Borders.R
\name{addbord}
\alias{addbord}
\title{Add borders to matrix}
\usage{
addbord(mat, int = 1, size = 1, outside = T)
}
\arguments{
\item{mat}{An input (image) matrix}

\item{int}{The intger to fill the matrix with}

\item{size}{Border thickness}

\item{outside}{Shall borders be drawn outside or inside the matrix ?}
}
\value{
A matrix with added borders
}
\description{
Add borders to matrix
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
