% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoNormalize.R
\name{autoNormalize}
\alias{autoNormalize}
\title{Image normalization with automated histogram adjustment}
\usage{
autoNormalize(x, inputRange, autoRange = T, RangeOnly = F,
  separate = T, step = 5 * 10^-3)
}
\arguments{
\item{x}{An array, or a 2D/3D matrix}

\item{inputRange}{The targeted min and max pixel intensities}

\item{autoRange}{A boolean indicating if inputRange should be calculated automatically}

\item{RangeOnly}{If set to true, only calculated ranges are returned}

\item{separate}{Shall image frames normalized separately ?}

\item{step}{A numeric value, indicating probs from quantile option for inputRange automated calculation}
}
\value{
An array, a matrix or an image with values ranged from 0 to 1
}
\description{
Normalize image pixel intensities between 0 and 1
}
\examples{
m = replicate(100, rnorm(1,mean=0,sd=0.5))
m.norm = autoNormalize(m,autoRange=F);m.autoNorm = autoNormalize(m,autoRange=T,step=10**-1)
plot(density(m));points(density(m.norm),type='l',col='red');points(density(m.autoNorm),type='l',col='blue')

}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
