% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_tpca_model_from_mean}
\alias{build_tpca_model_from_mean}
\title{Build TPCA model from mean shape}
\usage{
build_tpca_model_from_mean(qmean, alpha_t_array)
}
\arguments{
\item{qmean}{matrix containing the mean shape}

\item{alpha_t_array}{list of tangent vectors from the mean shape}
}
\value{
covdata list containing covariance matrix and the coefficients of eigen projections
}
\description{
Build TPCA model from mean shape
}
