% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphr.R
\name{buildCCMFromSpecificConfigurations}
\alias{buildCCMFromSpecificConfigurations}
\title{Convert configurations to CCM}
\usage{
buildCCMFromSpecificConfigurations(param_values, configurations)
}
\arguments{
\item{param_values}{A named list of vectors/lists. The names of the list are
the names of the parameters (columns) in the morphological field. The
vectors/lists contain the possible values that the parameter can have.
If \code{param_values} is NULL (default), then an empty field is created.
Use \code{editable = TRUE} in \code{\link{installMorphField}} to fill the
empty field in the browser.}

\item{configurations}{Optional. The configurations are a list of the valid
parameter configurations as alternative to the CCM. If the \code{ccm}
is also given, \code{configurations} are ignored. See details.}
}
\description{
If you specify your morphological field with \code{configurations},
then they can be converted to a CCM (cross-consistency matrix) using this
function.
}
