% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphr.R
\name{buildNestedListFromDataFrame}
\alias{buildNestedListFromDataFrame}
\title{Convert a long table format data.frame to a nested list}
\usage{
buildNestedListFromDataFrame(df)
}
\arguments{
\item{df}{A \code{data.frame} to be converted to a nested list.}
}
\value{
The converted data as nested list.
}
\description{
Use this function if you have a data.frame in long table format (with
redundant information in some columns) and need a nested list expected as
input to \code{\link{installMorphField}}, e.g. arguments \code{param_values},
\code{value_descriptions}, \code{specific_configurations}.
}
\details{
The redundant columns must be on the left, the non-redundant on the right
side of the data.frame.
}
