% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphr.R
\name{buildUnconstrainedCCM}
\alias{buildUnconstrainedCCM}
\title{Create a completely unconstrained CCM from parameter values}
\usage{
buildUnconstrainedCCM(param_values)
}
\arguments{
\item{param_values}{A named list of vectors/lists. The names of the list are
the names of the parameters (columns) in the morphological field. The
vectors/lists contain the possible values that the parameter can have.
If \code{param_values} is NULL (default), then an empty field is created.
Use \code{editable = TRUE} in \code{\link{installMorphField}} to fill the
empty field in the browser.}
}
\description{
All entries in the CCM (cross-consistency matrix) will be \code{TRUE}.
}
