% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphr.R
\name{ccmFromDataFrame}
\alias{ccmFromDataFrame}
\title{Convert data.frame to CCM}
\usage{
ccmFromDataFrame(df, param_values)
}
\arguments{
\item{df}{The CCM data.frame as returned by \code{\link{dataFrameFromCCM}}.}

\item{param_values}{A named list of vectors/lists. The names of the list are
the names of the parameters (columns) in the morphological field. The
vectors/lists contain the possible values that the parameter can have.
If \code{param_values} is NULL (default), then an empty field is created.
Use \code{editable = TRUE} in \code{\link{installMorphField}} to fill the
empty field in the browser.}
}
\description{
Use this function to convert the CCM from a tabular format used with \pkg{DT}
back to the original morphr format.
}
