% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_covariance}
\alias{compute_covariance}
\title{Compute covariance of a set of tangent vectors (alpha_t_array) at the mean shape (qmean)}
\usage{
compute_covariance(qmean, alpha_t_array)
}
\arguments{
\item{qmean}{matrix containing the mean shape}

\item{alpha_t_array}{list of tangent vectors from the mean shape}
}
\value{
covdata list containing covariance matrix and the coefficients of eigen projections
}
\description{
Compute covariance of a set of tangent vectors (alpha_t_array) at the mean shape (qmean)
}
