% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphr.R
\name{convertConfigsToExtended}
\alias{convertConfigsToExtended}
\title{Convert configurations to new extended format}
\usage{
convertConfigsToExtended(configurations)
}
\arguments{
\item{configurations}{The configurations in the old compact
format, see \code{\link{installMorphField}}.}
}
\value{
Configurations in the new extended format.
}
\description{
This function converts configurations in the old
compact format into the new extended format.
}
