% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphr.R
\name{convertSpecConfToExtended}
\alias{convertSpecConfToExtended}
\title{Convert specific configurations to new extended format}
\usage{
convertSpecConfToExtended(specific_configurations)
}
\arguments{
\item{specific_configurations}{The specific configurations in the old compact
format, see \code{\link{installMorphField}}.}
}
\value{
Specific configurations in the new extended format.
}
\description{
This function converts specific configurations (specifications) in the old
compact format into the new extended format.
}
