% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{curve_to_q}
\alias{curve_to_q}
\title{Convert coordinate curve to the q function}
\usage{
curve_to_q(p)
}
\arguments{
\item{p}{coordinate curve}
}
\value{
q function
}
\description{
Convert coordinate curve to the q function
}
