% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphr.R
\name{dataFrameFromCCM}
\alias{dataFrameFromCCM}
\title{Convert CCM to data.frame}
\usage{
dataFrameFromCCM(param_values, ccm)
}
\arguments{
\item{param_values}{A named list of vectors/lists. The names of the list are
the names of the parameters (columns) in the morphological field. The
vectors/lists contain the possible values that the parameter can have.
If \code{param_values} is NULL (default), then an empty field is created.
Use \code{editable = TRUE} in \code{\link{installMorphField}} to fill the
empty field in the browser.}

\item{ccm}{The CCM as returned by \code{\link{buildCCMFromSpecificConfigurations}}.}
}
\description{
Use this function to display the CCM in a tabular format, e.g. with \pkg{DT}.
}
