% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse}
\alias{ellipse}
\title{Compute the coordinates of points on an ellipse}
\usage{
ellipse(major, minor, alpha = 0, x = 0, y = 0, n = 100)
}
\arguments{
\item{major, minor}{length of the major and minor axis}

\item{alpha}{angle, in trigonometric convention (from the horizontal, positive counter-clockwise, in radians).}

\item{x, y}{coordinates of the center of the ellipse (around which the rotation is computed if alpha is not 0).}

\item{n}{number of points to compute on the ellipse.}
}
\description{
Compute the coordinates of points on an ellipse
}
\examples{
plot(ellipse(10, 5), asp=1)
plot(ellipse(10, 5, alpha=pi/4), asp=1)
plot(ellipse(10, 5, alpha=pi/4, x=3, y=4, n=200), asp=1)
}
