% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_func.R
\name{form_basis_d}
\alias{form_basis_d}
\title{Forms the basis for the tangent space of diffeomorphisms
This is actually an L2 [0,1] space}
\usage{
form_basis_d(d, T_col)
}
\arguments{
\item{d}{number Fourier basis elements, d for sin(kt), cos(kt), 0<=k<=d-1}

\item{T_col}{dimension of sample points}
}
\value{
matrix of the basis
}
\description{
Forms the basis for the tangent space of diffeomorphisms
This is actually an L2 [0,1] space
}
