% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_func.R
\name{form_basis_l2_r3}
\alias{form_basis_l2_r3}
\title{Form the Schauder basis for the tangent space L2(R3)}
\usage{
form_basis_l2_r3(d, T_col)
}
\arguments{
\item{d}{number Fourier basis elements, d for sin(kt), cos(kt), 0<=k<=d-1}

\item{T_col}{dimension of sample points}
}
\value{
matrix of the basis elements of length 6*d
}
\description{
Form the Schauder basis for the tangent space L2(R3)
}
