% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_func.R
\name{form_basis_o_q}
\alias{form_basis_o_q}
\title{Form the tangent space basis of the quotient space of Q/SO(n)}
\usage{
form_basis_o_q(B, q)
}
\arguments{
\item{B}{the basis for L2 (Rn)}

\item{q}{shape}
}
\value{
list of basis elements of T_q(Q/SO(n))
}
\description{
Form the tangent space basis of the quotient space of Q/SO(n)
}
