% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_func.R
\name{form_basis_of_tangent_space_of_s_at_q}
\alias{form_basis_of_tangent_space_of_s_at_q}
\title{Form the tangent space basis (T_q(S)) of the quotient space  S = C/(D_q x O_q)}
\usage{
form_basis_of_tangent_space_of_s_at_q(B, G_O_q)
}
\arguments{
\item{B}{Orthogonalized basis for L2 (Rn)}

\item{G_O_q}{The basis for the tangent space T_q(O_q) (using \code{\link{form_basis_o_q}}}
}
\value{
list of basis elements of T_q(S)
}
\description{
Form the tangent space basis (T_q(S)) of the quotient space  S = C/(D_q x O_q)
}
