% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_func.R
\name{geodesic_flow}
\alias{geodesic_flow}
\title{Compute the geodesic flow at q in the direction of w for an unit time}
\usage{
geodesic_flow(q, w, stp)
}
\arguments{
\item{q}{shape}

\item{w}{tangent vector at q}

\item{stp}{step size}
}
\value{
list of the shapes along the geodesic flow
}
\description{
Compute the geodesic flow at q in the direction of w for an unit time
}
