% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconstruction.R
\name{geodilate}
\alias{geodilate}
\title{Geomorphometric operations : geodilation}
\usage{
geodilate(marker, reference, element)
}
\arguments{
\item{marker}{An image matrix marker}

\item{reference}{An image matrix reference}

\item{element}{Structural element}
}
\value{
The geodesic dilation of the image
}
\description{
Geodilation
}
\author{
Marion Leduc, \email{marion.leduc@gustaveroussy.fr}
}
