% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconstruction.R
\name{geoerode}
\alias{geoerode}
\title{Geomorphometric operations : geoerosion}
\usage{
geoerode(marker, reference, element)
}
\arguments{
\item{marker}{An image matrix marker}

\item{reference}{An image matrix reference}

\item{element}{Structural element}
}
\value{
The geodesic erosion of the image
}
\description{
Geoerosion
}
\author{
Marion Leduc, \email{marion.leduc@gustaveroussy.fr}
}
