% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_moments.R
\name{get_moment}
\alias{get_moment}
\alias{gm}
\title{Get an image moment}
\usage{
get_moment(x, i = 0, j = 0)

gm(x, i = 0, j = 0)
}
\arguments{
\item{x}{a moments matrix generated by of the \code{\link[=img_moments]{img_moments()}} functions}

\item{i, j}{the 0-based indexes of the matrix}
}
\description{
Extract an image moment from a square moments matrix using 0-based indexing.
}
\examples{
x <- img_read(system.file("extdata", "plank/16199658.jpg", package="morphr"))
X_binary <- x < 1
m <- img_moments(X_binary, 3)
m
get_moment(m, 0,0)
gm(m, 0,0)
}
