% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{get_plot_output_list}
\alias{get_plot_output_list}
\title{Utility function for shiny app}
\usage{
get_plot_output_list(max_plots, input_n, X)
}
\arguments{
\item{input_n}{number of iterations}

\item{X}{shape}

\item{maxplots}{maximum number of plots}
}
\value{
plots
}
\description{
Utility function for shiny app
}
