% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggimg_grid.R
\name{ggimg_grid}
\alias{ggimg_grid}
\title{Plot a grid of images}
\usage{
ggimg_grid(imgs, scale = 0.001)
}
\arguments{
\item{imgs}{vector of paths to images}
}
\description{
Plot a grid of images
}
\examples{
paths <- list.files(
  system.file("extdata", "plank", package="morphr"),
  full.names=TRUE
)
ggimg_grid(imgs=sample(paths, 15), scale=0.002)
ggimg_grid(imgs=sample(paths, 50), scale=0.001)
}
