% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_func.R
\name{gram_schmidt}
\alias{gram_schmidt}
\title{Orthogonalized the vectors X}
\usage{
gram_schmidt(X)
}
\arguments{
\item{X}{list of vectors}
}
\value{
orthogonalized vectors
}
\description{
Orthogonalized the vectors X
}
