% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{HighPass}
\alias{HighPass}
\title{Digital filtering : High Pass filter}
\usage{
HighPass(x, ...)
}
\arguments{
\item{x}{An image matrix}
}
\value{
The convolved image
}
\description{
High Pass filter
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
