% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_adjust_gamma.R
\name{img_adjust_gamma}
\alias{img_adjust_gamma}
\title{Perform gamma correction}
\usage{
img_adjust_gamma(x, gamma = 1)
}
\arguments{
\item{x}{an image, of type \code{\link[imager:cimg]{imager::cimg()}}.}

\item{gamma}{gamma value, in [0,+Inf[ where 1 means no change, <1 make the image lighter, >1 makes the image darker.}
}
\description{
Transorm the input image according to the equation x^gamma.
}
\examples{
x <- img_read(system.file("extdata", "amphipoda/33463695.jpg",
             package="morphr")) \%>\% img_show()
img_adjust_gamma(x, 0.2) \%>\% img_show()
img_adjust_gamma(x, 2) \%>\% img_show()
}
