% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_center.R
\name{img_center}
\alias{img_center}
\title{Center an object in an image}
\usage{
img_center(x, w, h, around = NULL, col = NULL)
}
\arguments{
\item{x}{an image, of type \code{\link[imager:cimg]{imager::cimg()}}.}

\item{w, h}{the width and height of the output image.}

\item{around}{a vector with the x,y coordinate of the point around which to center; by default the centroid of the input image.}

\item{col}{the colour to use for padding; by default the median colour of the border.}
}
\description{
Center the input object within an output image of given dimensions, around a set of coordinates (by default the centroid of the object).
}
\examples{
x <- 1 - img_read(system.file("extdata", "amphipoda/33463695.jpg", package="morphr"))
img_show(x)
# center with all defaults
# = around the center of mass of grey levels
xc <- img_center(x, 350, 350) \%>\% img_show()
# check that the new centroid is around the middle of the image [175,175]
img_centroid(xc)

# center around the centroid of the binary image
# = does not take grey levels into account
xc <- img_center(x, 350, 350, around=img_centroid(x>0)) \%>\% img_show()
img_centroid(xc>0)
# the binary centroid is in the middle
img_centroid(xc)
# but the grey levels center of mass is elsewhere
}
