% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_centroid.R
\name{img_centroid}
\alias{img_centroid}
\title{Centroid of object on image}
\usage{
img_centroid(x)
}
\arguments{
\item{x}{a greyscale image, of type \code{\link[imager:cimg]{imager::cimg()}}}
}
\value{
A vector with the x and y coordinates of the centroid.
}
\description{
Compute the coordinates of the centroid of the object based on image moments. The centroid is based on shape if the image is binary and is weighted by the pixel values if it is not (in which case it is usually called the centre of mass).
}
\examples{
x <- img_read(system.file("extdata", "plank/16199658.jpg", package="morphr"))
X_binary <- x < 1
X_inverted <- 1 - x
img_centroid(X_binary)
img_centroid(X_inverted)
}
