% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_chop.R
\name{img_chop}
\alias{img_chop}
\title{Chop away borders from an image}
\usage{
img_chop(x, top = 0, right = 0, bottom = 0, left = 0)
}
\arguments{
\item{x}{input image (a \code{\link[imager:cimg]{imager::cimg()}} object).}

\item{top}{number of pixels to chop from the top.}

\item{right}{same, from the right.}

\item{bottom}{same, from the bottom}

\item{left}{same, from the left.}
}
\value{
The image with the pixels chopped off.
}
\description{
Chop away borders from an image
}
\examples{
x <- img_read(system.file("extdata", "blob.jpg", package="morphr"))
img_show(x)
# chop off some pieces around the main object
img_chop(x, t=17, r=15, b=50, l=15) \%>\% img_show()
# just remove the legend
img_chop(x, b=31) \%>\% img_show()
}
