% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_guess_background.R
\name{img_guess_background}
\alias{img_guess_background}
\title{Guess the background colour of an image}
\usage{
img_guess_background(x)
}
\arguments{
\item{x}{an image, of type \code{\link[imager:cimg]{imager::cimg()}}.}
}
\description{
Guess the background colour as the dominant colour at the border of the image.
}
\examples{
x <- img_read(system.file("extdata", "16195419.jpg", package="morphr")) \%>\% img_show()
img_guess_background(x)
x <- img_read(system.file("extdata", "shape.png", package="morphr")) \%>\% img_show()
img_guess_background(x)
}
