% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_invert.R
\name{img_invert}
\alias{img_invert}
\title{Invert a greyscale image}
\usage{
img_invert(x)
}
\arguments{
\item{x}{an image, of type \code{\link[imager:cimg]{imager::cimg()}}.}
}
\value{
The inverted image, as a \code{\link[imager:cimg]{imager::cimg()}} object.
}
\description{
This is a simple 1-x operation but having it as a function ease its inclusing in within a piped chain of commands
}
\examples{
x <- img_read(system.file("extdata", "blob.jpg", package="morphr"))
img_show(x)
img_show(img_invert(x))
}
