% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_make_horizontal.R
\name{img_make_horizontal}
\alias{img_make_horizontal}
\title{Rotate object on image horizontally}
\usage{
img_make_horizontal(x)
}
\arguments{
\item{x}{an image, of type \code{\link[imager:cimg]{imager::cimg()}}.}
}
\description{
Measure the orientation angle with \code{\link[=img_orientation]{img_orientation()}}, rotate the image, then use \code{\link[=img_centroid]{img_centroid()}} to detect the centroid and flip the image so that the centroid is is the top-right quadrant. All this works with a light-on-dark image.
}
\examples{
x <- img_read(system.file("extdata", "amphipoda/33463695.jpg",
             package="morphr"), invert=TRUE) \%>\% img_show()
img_make_horizontal(x) \%>\% img_show()
}
