% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_make_transparent.R
\name{img_make_transparent}
\alias{img_make_transparent}
\title{Make a greyscale image transparent}
\usage{
img_make_transparent(x)
}
\arguments{
\item{x}{input greyscale image, of type \code{\link[imager:cimg]{imager::cimg()}}.}
}
\description{
Turn white into fully transparent and black into fully opaque in a greyscale image.
}
\examples{
x <- img_read(system.file("extdata", "16195419.jpg", package="morphr"))
img_show(x)
img_show(img_make_transparent(x))
}
