% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_read.R
\name{img_read}
\alias{img_read}
\title{Read a greyscale image}
\usage{
img_read(file)
}
\arguments{
\item{file}{path to the input image file.}
}
\value{
A \code{\link[imager:cimg]{imager::cimg()}} object, which is a four-dimensional numeric array containing x,y,z,c coordinates, where x,y are dimensions, z is depth, and c is colour.
}
\description{
Read a greyscale image
}
\examples{
path <- system.file("extdata", "16195419.jpg", package="morphr")
x <- img_read(path)
img_show(x)
}
