% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_write.R
\name{img_write}
\alias{img_write}
\title{Write a grayscale image}
\usage{
img_write(x, file, quality = 0.9)
}
\arguments{
\item{x}{image data matrix: a 2D array of values in \verb{[0,1]}.}

\item{file}{path to the output file; the extension determines the format.}

\item{quality}{image quality, for JPG output only.}
}
\value{
The path to the written file, invisibly (assign it to a variable to capture it).
}
\description{
Write a grayscale image
}
\examples{
path <- system.file("extdata", "16195419.jpg", package="morphr")
x <- img_read(path)
out <- img_write(x, file=tempfile(fileext=".jpg"))
out
out <- img_write(x, file=tempfile(fileext=".png"))
out
}
