% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imread.R
\name{imread}
\alias{imread}
\title{Read an image}
\usage{
imread(path)
}
\arguments{
\item{path}{path to the input image file.}
}
\value{
A matrix with the pixel values of the image.
}
\description{
Read an image
}
\examples{
img_file  <- system.file("extdata", "16195419.jpg", package="morphr")
x <- imread(img_file)
imshow(x)
}
