% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_py_deps.R
\name{install_py_deps}
\alias{install_py_deps}
\title{Install python dependencies}
\usage{
install_py_deps()
}
\description{
Tests whether os, numpy and skimage are available. Install them if they are not.
}
\examples{
\dontrun{
# check which python version is detected and install python modules
library("reticulate")
py_config()
install_py_deps()

# use python3 (works only in a new R session)
library("reticulate")
use_python(system("which python3", intern=TRUE))
py_config()
install_py_deps()
}
}
