% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{Lap}
\alias{Lap}
\title{Digital filtering : Laplacian filter}
\usage{
Lap(x, choice = 1, ...)
}
\arguments{
\item{x}{An image matrix}

\item{choice}{What discrete kernel shall be used if several available#'}

\item{direction}{direction for Sobel edges detection. Can be 'left', 'right', or 'both'}
}
\value{
The convolved image
}
\description{
Laplacian filter
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
