% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{LowPass}
\alias{LowPass}
\title{Digital filtering : Low Pass filter}
\usage{
LowPass(x, ...)
}
\arguments{
\item{x}{An image matrix}
}
\value{
The convolved image
}
\description{
Low Pass filter
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
