% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transparent.R
\name{make_transparent}
\alias{make_transparent}
\title{Make a greyscale image transparent}
\usage{
make_transparent(x)
}
\arguments{
\item{x}{matrix representing a grayscale image (in [0,1] or [0,255], where the maximum is white).}
}
\description{
Turn white into fully transparent and black into fully opaque in a greyscale image.
}
\examples{
# create an image
amph <- list.files(
          system.file("extdata", "amphipoda", package="morphr"),
          full.names=TRUE
       )
img <- morph(amph)
dim(img)

# add transparency = turn it into RGBA
imga <- make_transparent(img)
dim(imga)

# show the difference between the two
grid::grid.newpage()
grid::grid.rect(gp=gpar(fill="dodgerblue"))
grid::grid.raster(img/255)

grid::grid.newpage()
grid::grid.rect(gp=gpar(fill="dodgerblue"))
grid::grid.raster(imga)
}
