% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho_space.R
\name{morpho_space}
\alias{morpho_space}
\title{Build morphological space}
\usage{
morpho_space(x, weights = NULL, ...)
}
\arguments{
\item{x}{data.frame or matrix of morphological descriptors. Should be all numeric and as close to normally-distributed as possible.}

\item{weights}{vector of objects weights, with no weights by default. Should have as many elements as \code{x} has rows but does not need to sum to 1.}

\item{...}{passed to \code{\link[FactoMineR:PCA]{FactoMineR::PCA()}}.}
}
\value{
An object of class \code{PCA}, see \code{\link[FactoMineR:PCA]{FactoMineR::PCA()}} for more information.
}
\description{
Build a morphological space by computing a Principal Component Analysis on a table of morphological descriptors.
}
\examples{
s <- morpho_space(plank[,-(1:2)], weights=plank$conc)
plot(s, choix="ind")
plot(s, choix="var")
}
