% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphospace.R
\name{morphospace}
\alias{morphospace}
\title{Build morphological space}
\usage{
morphospace(x, weights = NULL, method = "PCA", ...)
}
\arguments{
\item{x}{data.frame or matrix of morphological descriptors. Should be all numeric and as close to normally-distributed as possible.}

\item{weights}{vector of objects weights, with no weights by default. Should have as many elements as rows in \code{x} but they do not need to sum to 1.}

\item{method}{dimensionality reduction method. Currently only "PCA" is implemented (with \code{\link[FactoMineR:PCA]{FactoMineR::PCA()}}).}

\item{...}{passed to the dimensionality reduction function.}
}
\value{
An object of class \code{morphospace}.
}
\description{
Build a morphological space by performing dimensionality reduction on a table of morphological descriptors.
}
\examples{
s <- morphospace(plank[,-(1:2)], weights=plank$conc)
plot(s, choix="ind")
plot(s, choix="var")
}
