\name{morphr-package}
\alias{morphr-package}
\alias{morphr}
\docType{package}
\title{
\packageTitle{morphr}
}
\description{
\packageDescription{morphr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{morphr}
\packageIndices{morphr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{morphr}

Maintainer: \packageMaintainer{morphr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
