% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reconstruction.R
\name{MorphRecons}
\alias{MorphRecons}
\title{Geomorphometric operations : reconstruction}
\usage{
MorphRecons(f, marker, reference, element, oc = NULL)
}
\arguments{
\item{f}{the function to be applied}

\item{marker}{An image matrix marker}

\item{reference}{An image matrix reference}

\item{element}{Structural element}

\item{oc}{Can be NULL, 'open' or 'close'; default is NULL}
}
\value{
The geodesic erosion of the image
}
\description{
Image reconstruction after geomorphometric operation
}
\details{
if oc is set to \code{\link{open}}, then \code{\link{geodilate}} shall be used as the working function;
if oc is set to \code{\link{close}}, then \code{\link{geoerode}} shall be used as the working function;
}
\author{
Marion Leduc, \email{marion.leduc@gustaveroussy.fr}
}
