% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pcacov}
\alias{pcacov}
\title{Compute PCA of matrix of observations}
\usage{
pcacov(C)
}
\arguments{
\item{C}{matrix of observations}
}
\value{
list of principal coefficients, latent (vector of eigenvalues of C), explained (percent of total variance)
}
\description{
Compute PCA of matrix of observations
}
