% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphr.R
\docType{data}
\name{plank}
\alias{plank}
\title{Morphology of planktonic organisms}
\format{A data.frame of 500 rows and 10 variables
\describe{
\item{id}{object identifier, which links to the corresponding image in the package data.}
\item{w}{concentration in situ, which is to be considered as weights for the morphological space construction.}
\item{area:mean}{morphological descriptors of each organism: area, major dimension length, minor dimension length (in pixels), alongation (=major/minor), fractal characteristic of the perimeter, circularity, average grey level, and coefficient of variation of the grey levels.}
}}
\source{
Underwater Vision Profiler samples from the GreenEdge 2016 cruise in Baffin Bay.
}
\usage{
plank
}
\description{
A dataset containing link to images, concentration, and morphological descriptors or planktonic organisms.
}
\keyword{datasets}
