% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_pca_variation}
\alias{plot_pca_variation}
\title{Display variation of pca}
\usage{
plot_pca_variation(pathname)
}
\arguments{
\item{pathname}{location of the file}

\item{alpha_t_array}{list of tangent vectors from the mean shape}
}
\description{
Display variation of pca
}
