% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeo_johnson.R
\name{predict.yeo_johnson}
\alias{predict.yeo_johnson}
\title{Transform new data using an already fitted Yeo-Johnson object}
\usage{
\method{predict}{yeo_johnson}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class 'yeo_johnson'}

\item{newdata}{a numeric vector of data to transform.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Transform new data using an already fitted Yeo-Johnson object
}
\examples{
# fit the Yeo=-Johnson transfomration on non-normal data
x <- rgamma(100, 1, 1)
hist(x)
xt <- yeo_johnson(x)
# apply the same transformation to new data
x2 <- rgamma(100, 1, 1)
hist(x2)
hist(predict(xt, newdata=x2))
}
