% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphr-shiny.R
\name{preselectMorphField}
\alias{preselectMorphField}
\title{Select cells in the morph field programmatically}
\usage{
preselectMorphField(proxy, param_values, pre_selections)
}
\arguments{
\item{proxy}{The \code{dataTableProxy()}.}

\item{param_values}{A named list of vectors/lists. The names of the list are
the names of the parameters (columns) in the morphological field. The
vectors/lists contain the possible values that the parameter can have.
If \code{param_values} is NULL (default), then an empty field is created.
Use \code{editable = TRUE} in \code{\link{installMorphField}} to fill the
empty field in the browser.}

\item{pre_selections}{Optional. Specify which cells shall be selected when the
app starts. \code{pre_selections} is a list of named lists with names
\code{param} and \code{value}.}
}
\description{
With this function, you can set certain cells as selected, e.g. when the app
starts.
}
