% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_func.R
\name{project_basis_tangent_space_of_c_at_q}
\alias{project_basis_tangent_space_of_c_at_q}
\title{Project the basis B on the tangent space T_q(C)}
\usage{
project_basis_tangent_space_of_c_at_q(B, q)
}
\arguments{
\item{B}{the basis for L2 (Rn)}

\item{q}{shape}
}
\value{
list of basis elements of T_q(C)
}
\description{
Project the basis B on the tangent space T_q(C)
}
