% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_func.R
\name{project_curve}
\alias{project_curve}
\title{Project the shape q to the space of closed curves, C}
\usage{
project_curve(q)
}
\arguments{
\item{q}{shape}
}
\value{
qnew, projected shape on C
}
\description{
Project the shape q to the space of closed curves, C
}
