% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_func.R
\name{project_to_basis}
\alias{project_to_basis}
\title{Project the set of vectors alpha_t_array into a basis set Y}
\usage{
project_to_basis(alpha_t_array, Y)
}
\arguments{
\item{alpha_t_array}{set of tangent vectors}

\item{Y}{a basis set}
}
\value{
list of projected vectors (X) and the coefficients of projections (X_proj)
}
\description{
Project the set of vectors alpha_t_array into a basis set Y
}
