% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_func.R, R/q_to_curve.R
\name{q_to_curve}
\alias{q_to_curve}
\alias{q_to_curve}
\title{Convert q to curve}
\usage{
q_to_curve(q)

q_to_curve(q)
}
\arguments{
\item{q}{vector of coordinates of curves}

\item{q}{vector of coordinates of curves}
}
\value{
numeric vector

numeric vector
}
\description{
Convert q to curve

Convert q to curve
}
