% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_svg_file}
\alias{read_svg_file}
\title{Read shape coordinates from svg files}
\usage{
read_svg_file(file_path, N = 100, r = 3)
}
\arguments{
\item{file_path}{location of svg files}

\item{N}{number of points (default value 100)}

\item{r}{number of iterations (default value 3)}
}
\value{
coordinates of shape
}
\description{
Read shape coordinates from svg files
}
