% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_ucf_multiple_levels}
\alias{read_ucf_multiple_levels}
\title{Read shape coordinates from ucf files}
\usage{
read_ucf_multiple_levels(filepath)
}
\arguments{
\item{file_path}{location of ucf files}
}
\value{
list of coordinates of shape
}
\description{
Read shape coordinates from ucf files
}
