% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segcomp.R
\name{ReadSeg}
\alias{ReadSeg}
\title{Restore compressed mask}
\usage{
ReadSeg(fi, asRDS = FALSE)
}
\arguments{
\item{fi}{File path}

\item{asRDS}{Was file saved as RDS?}
}
\value{
A matrix object
}
\description{
Restore compressed mask
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
