% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repose_curve}
\alias{repose_curve}
\title{Scale, translate and rotate curve p}
\usage{
repose_curve(p, L, R, POS)
}
\arguments{
\item{p}{curve}

\item{L}{scaling factor}

\item{R}{rotation matrix}

\item{POS}{translation}
}
\value{
scaled, translated and rotated version of curve p
}
\description{
Scale, translate and rotate curve p
}
