% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize_pad.R
\name{resize_pad}
\alias{resize_pad}
\title{Resize image by padding or cropping.}
\usage{
resize_pad(x, target_size, value)
}
\arguments{
\item{x}{An input (image) matrix}

\item{target_size}{The target size of the resulting image matrix}

\item{value}{The pixel value to be added. If absent, the median from the original image will be used}
}
\value{
A centered resized matrix
}
\description{
Resize image by padding or cropping.
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
