% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_eigen_projections}
\alias{save_eigen_projections}
\title{Save the eigen projections of the pca}
\usage{
save_eigen_projections(covdata, alpha_t_array, eigendirs, filename)
}
\arguments{
\item{covdata}{list of covariance matrix, eigen values returned from \code{\link{compute_covariance}}}

\item{alpha_t_array}{list of tangent vectors from the mean shape}

\item{eigendirs}{the principal eigen directions}

\item{filename}{exported filename}
}
\value{
csv of eigen projections
}
\description{
Save the eigen projections of the pca
}
