% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segcomp.R
\name{SaveSeg}
\alias{SaveSeg}
\title{Save mask using rle algorithm}
\usage{
SaveSeg(Mask, ExportDir, FileName, asRDS = FALSE)
}
\arguments{
\item{Mask}{The image mask to be saved}

\item{asRDS}{Shall file be saved as RDS?}

\item{ExportName}{The file saving location}
}
\description{
Save mask using rle algorithm
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
