% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{Sharp}
\alias{Sharp}
\title{Digital filtering : Sharpen}
\usage{
Sharp(x, ...)
}
\arguments{
\item{x}{An image matrix}

\item{direction}{direction for Sobel edges detection. Can be 'left', 'right', or 'both'}

\item{choice}{What discrete kernel shall be used if several available#'}
}
\value{
The convolved image
}
\description{
Sharpen
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
