% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoNormalize.R
\name{sigmoNormalize}
\alias{sigmoNormalize}
\title{Sigmoidal image normalization}
\usage{
sigmoNormalize(x, lambda = 2, z0 = 1.5, scaled = F)
}
\arguments{
\item{x}{a 2D matrix}

\item{lambda}{smoothness of transformation}

\item{z0}{center of sigmoidal transformation}

\item{scaled}{are input values scaled ?}
}
\value{
Normalized 2D matrix, with values ranging from 0 to 1
}
\description{
Flatten image extrema by sigmoidal transformation
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
