% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_func.R
\name{smooth}
\alias{smooth}
\title{Smooth a function based on degree of smoothing}
\usage{
smooth(y, span)
}
\arguments{
\item{y}{function to be smoothed}

\item{span}{degree of smoothing}
}
\value{
return a smoothed function
}
\description{
Smooth a function based on degree of smoothing
}
