% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{Sobel}
\alias{Sobel}
\title{Digital filtering : Edges detection}
\usage{
Sobel(x, direction = "both", ...)
}
\arguments{
\item{x}{An image matrix}

\item{direction}{direction for Sobel edges detection. Can be 'left', 'right', or 'both'}
}
\value{
The convolved image
}
\description{
Sobel filter
}
\author{
Allan Sauvat, \email{allan.sauvat@gustaveroussy.fr}
}
