% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphr.R
\name{sortConfigs}
\alias{sortConfigs}
\title{Sort configurations that are in the new format}
\usage{
sortConfigs(configurations)
}
\arguments{
\item{The}{configurations in the new exptended format, see
\code{\link{installMorphField}}.}
}
\description{
Sorts configuratios.
}
