% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_py_deps.R
\name{test_module}
\alias{test_module}
\title{Test whether a python module is installed}
\usage{
test_module(module, import_name = module)
}
\arguments{
\item{module}{name of the module to test}

\item{import_name}{name under which the module should be imported (when different from the module name); for e.g. scikit-image which is imported as skimage}
}
\description{
Test whether a python module is installed
}
