% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trapz}
\alias{trapz}
\title{Numerical integration of the function y(x)}
\usage{
trapz(x, y, dims = 1)
}
\arguments{
\item{x}{the set of points where y is computed on}

\item{y}{the integrand}
}
\value{
the result
}
\description{
Numerical integration of the function y(x)
}
