% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{verify_shapes}
\alias{verify_shapes}
\title{Display shape orientation}
\usage{
verify_shapes(pathname)
}
\arguments{
\item{pathname}{location of the ucf file containing hape}
}
\description{
Display shape orientation
}
