% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeojohnson.R
\name{yeojohnson}
\alias{yeojohnson}
\title{Yeo-Johnson transformation}
\usage{
yeojohnson(x, lambda = NULL, eps = 0.001)
}
\arguments{
\item{x}{a numeric vector of values to make normal.}

\item{lambda}{the value of the lambda parameter. When NULL, an appropriate value is estimated from the data.}

\item{eps}{the tolerance value under which lamba is considered to be 0 and that is used to choose the appropriate formula.}
}
\value{
A list of class \code{yeojohnson} with elements
\item{data}{original data}
\item{transformed}{transformed data}
\item{lambda}{lambda value used for the transformation (possibly estimated)}
}
\description{
Perform Yeo-Johnson transformation to attempt normalization.
}
\details{
The Yeo-Johnson is similar to the Box-Cox method, however it allows for the transformation of non-positive data as well.
}
\examples{
# simulate non-normal data
x <- rgamma(100, 1, 1)
hist(x)
xt <- yeojohnson(x)
hist(xt$transformed)
}
\references{
Yeo, I. K., & Johnson, R. A. (2000). A new family of power transformations to improve normality or symmetry. Biometrika.
}
\seealso{
\code{\link[=predict.yeojohnson]{predict.yeojohnson()}} to apply the same transformation to a new dataset.
}
