% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggimg_grid.R
\name{ggimg_grid}
\alias{ggimg_grid}
\title{Plot a grid of images}
\usage{
ggimg_grid(imgs, scale = 0.001, fun = NULL, ...)
}
\arguments{
\item{imgs}{vector of paths to images}

\item{scale}{scaling factor used to display the images; from pixels to plot dimensions (in [0,1]).}

\item{fun}{function to apply to each image, as a pre-processing step (a typical one is \code{\link[=img_chop]{img_chop()}} to remove a scale bar added to the image).}

\item{...}{passed to \code{fun}.}
}
\description{
Plot a grid of images
}
\examples{
paths <- list.files(system.file("extdata", "plank",
                    package="morphr"), full.names=TRUE)
ggimg_grid(imgs=sample(paths, 15), scale=0.002)
ggimg_grid(imgs=sample(paths, 56), scale=0.001)
}
