% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmorph.R
\name{ggmorph_radial}
\alias{ggmorph_radial}
\title{Display morphs radially within a morphological space}
\usage{
ggmorph_radial(space, imgs, dimensions = c(1, 2), directions = 4,
  steps = 2, n_imgs = 5, adjust_grey = TRUE, scale = 0.01)
}
\arguments{
\item{space}{morphological space object created by [morpho_space()].}

\item{imgs}{vector of paths to images; should have as many elements as are in the morphological space}

\item{dimensions}{couple of dimensions to plot.}

\item{directions}{number of radial directions along which to display morphs.}

\item{steps}{number of steps in each direction at which to display a morph (a morph is always displayed in the middle of the space).}

\item{n_imgs}{number of images to randomly select and morph in each tile.}

\item{adjust_grey}{whether to adjust the mean grey level of the morphed image to match that of the orignal images it was constructed with. Usually, computing the average(i.e. the morphing) results in lighter images; this compensates it a little. Note that it requires more intense computatin and is therefore switched off by default.}

\item{scale}{scaling factor used to display the images; from pixels to morphological space dimensions units.}
}
\description{
Display morphs radially within a morphological space
}
\examples{
space <- morpho_space(plank[,-(1:2)], weights=plank$conc)
img_root <- system.file("extdata", "plank", package="morphr")
imgs <- file.path(img_root, paste0(plank$id, ".jpg"))
set.seed(1)
ggmorph_radial(space, imgs)
ggmorph_radial(space, imgs, directions=10, steps=4, scale=0.005)
ggmorph_radial(space, imgs, dim=c(3,4), scale=0.004)
}
