% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imread.R
\name{img_read}
\alias{img_read}
\title{Read a grayscale image}
\usage{
img_read(file)
}
\arguments{
\item{file}{path to the input image file.}
}
\value{
A matrix with the pixel values of the image. Indexes are counted from the top left (i.e. like a matrix).
}
\description{
Read a grayscale image
}
\examples{
path <- system.file("extdata", "16195419.jpg", package="morphr")
x <- img_read(path)
img_show(x)
}
