% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_show.R
\name{img_show}
\alias{img_show}
\title{Display an array of pixel as a greyscale image}
\usage{
img_show(x)
}
\arguments{
\item{x}{a greyscale image, with grey levels coded in \verb{[0, 1]}. This can be either a 2D matrix of pixels or a 4D \code{\link[imager:cimg]{imager::cimg()}} object.}
}
\value{
The (x, y, z) list used to plot the image, which is also suitable for contour, persp, etc. returned invisibly.
}
\description{
Display an array of pixel as a greyscale image
}
\examples{
# check that the overall grey level is reflected in the plot
img_show(matrix(runif(100, 0, 0.5), nrow=10))
img_show(matrix(runif(100, 0.5, 1), nrow=10))

# display an image
path <- system.file("extdata", "16195419.jpg", package="morphr")
x <- img_read(path)
img_show(x)
}
