% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imshow.R
\name{imshow}
\alias{imshow}
\title{Display an array of pixel as a greyscale image}
\usage{
imshow(x)
}
\arguments{
\item{x}{input array of pixels, with grey levels coded in [0, 255]}
}
\value{
The (x, y, z) list used to plot the image, which is also suitable for contour, persp, etc. returned invisibly.
}
\description{
Display an array of pixel as a greyscale image
}
\examples{
# check that the overall grey level is reflected in the plot
imshow(matrix(runif(100, 0, 100), nrow=10))
imshow(matrix(runif(100, 150, 250), nrow=10))

# read an image through python
io <- reticulate::import("skimage.io", as="io")
img <- io$imread(
  system.file("extdata", "amphipoda", "33463695.jpg", package="morphr"),
  as_grey=TRUE
)
# display it
imshow(img)
}
