% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph.R
\name{morph}
\alias{morph}
\title{Morph several images into one}
\usage{
morph(paths, dest = "")
}
\arguments{
\item{paths}{paths to image files to morph together}

\item{dest}{path to a new file in which to write the morphed image. When empty (the default), the morphed image is returned as a matrix. When not empty, the image is written to disk and returned invisibly by the function.}
}
\value{
A matrix containing the pixel values for the morphed image, in [0,255] (black to white).
}
\description{
Morph several images into one
}
\examples{
source <- system.file("extdata", "amphipoda", package="morph")
files <- list.files(source, full.names=TRUE)
img <- morph(files)
image(img, col=grey(1:255/255), asp=1)
\dontrun{
img <- morph(files, dest=path.expand("~/amphipoda_morph.jpg"))
}
}
