% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_dir.R
\name{morph_dir}
\alias{morph_dir}
\title{Morph images from a directory into one}
\usage{
morph_dir(path, n = 15, write = FALSE)
}
\arguments{
\item{path}{path to the directory containing image files to morph together}

\item{n}{how many images to randomly select within the directory. Set this to a very high number to process all}

\item{write}{boolean, whether to write the image next to the directiry in which the files are read. When FALSE (the default), the morphed image is returned, as a matrix. When TRUE, the image is written to disk and returned invisibly by the function.}
}
\value{
A matrix containing the pixel values for the morphed image, in [0,255] (black to white).
}
\description{
Morph images from a directory into one
}
\examples{
img <- morph_dir(system.file("extdata", "amphipoda", package="morph"))
image(img, col=grey(1:255/255), asp=1)
img <- morph_dir(system.file("extdata", "amphipoda", package="morph"), n=3)
image(img, col=grey(1:255/255), asp=1)
}
