% Please edit documentation in R/morphr.R
\name{morphfield}
\alias{morphfield}
\title{Create an HTML morphological field widget using the DataTables library}
\usage{
morphfield(param_values, specific_configurations = NULL)
}
\description{
This function creates an HTML widget to display a morphological field (a kind
of table where each column represents the possible configurations of a
certain parameter) using the JavaScript library DataTables.
}
