\name{morphr-package}
\alias{morphr-package}
\alias{morphr}
\docType{package}
\title{
\packageTitle{morphr}
}
\description{
\packageDescription{morphr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{morphr}
\packageIndices{morphr}

}
\author{
\packageAuthor{morphr}

Maintainer: \packageMaintainer{morphr}
}
\references{

}
\keyword{ package }
\seealso{
}
\examples{

}
