% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphr.R
\name{paramValuesToDataFrame}
\alias{paramValuesToDataFrame}
\title{Create a fixed dimension data.frame from variable length lists of morphological field parameter values}
\usage{
paramValuesToDataFrame(param_values, value_descriptions = NULL)
}
\arguments{
\item{param_values}{A named list of vectors/lists. The names of the list are
the names of the parameters (columns) in the morphological field. The
vectors/lists contain the possible values that the parameter can have.}

\item{value_descriptions}{Optional. Each parameter value in param_values can
have an accompanying (long) description that will be shown as tooltip/popover.
The structure of \code{value_descriptions} is like that of \code{param_values},
but instead of containing parameter values as list elements, the parameter
values must be names whose list elements are the description texts.}
}
\description{
A morphological field can consist of parameters with variable number of
values. However, a DataTable accepts only a data.frame, which must
have the same number of rows in each column. Therefore, this function takes
a list of lists (each list representing one list of parameter configurations,
i.e. one column) and fills the empty cells with empty character strings to
create a fixed dimension data.frame.
}
