% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_curve}
\alias{plot_curve}
\title{Plot the curve p with color value}
\usage{
plot_curve(p, colorstr, l = FALSE)
}
\arguments{
\item{p}{coordinates of curve}

\item{colorstr}{color value}

\item{l}{Default value is False, which means returning a plot}
}
\description{
Plot the curve p with color value
}
